eliminated generator input in GenerateAntScript
